00/python_server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Jeffermus",email = "70596979+Jeffermus@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
